update account_model to users_model
I was using account_model to have the user data model, but it would be confusing when I created the bank account data.
I am also creating the user account differently. I am implementing the hash for the password.
I also made changes to the database to accommodate the change, I created a specific table to only have hashed passwords, with a foreign_key for users